feat(converter): add tabs for unit categories
