+++ data/conversions/length.json
@@ -1,0 +1,10 @@
+{
+  "meter": 1,
+  "kilometer": 1000,
+  "centimeter": 0.01,
+  "millimeter": 0.001,
+  "mile": 1609.34,
+  "yard": 0.9144,
+  "foot": 0.3048,
+  "inch": 0.0254
+}
