+++ data/conversions/weight.json
@@ -1,0 +1,8 @@
+{
+  "kilogram": 1,
+  "gram": 0.001,
+  "milligram": 0.000001,
+  "pound": 0.45359237,
+  "ounce": 0.0283495,
+  "ton": 1000
+}
